+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog
 
+## [1.1.8]
+
+- Update toast for successful swap and add abort controller to fetch assets function.
+- Update the coinbase wallet id.
+
 ## [1.1.7]
 
 - Fixed USD value formatting and rate display.
 - Added Evm and Sui support for Backpack wallet.
+- Added Husky pre-commit hook to run ESLint with --fix, restage updated files, and block empty commits.
 
 ## [1.1.6]
 
@@ -19,7 +25,6 @@
 - Updated balance retrieval function to use fallback RPC clients as default instead of default wallet RPCs.
 - Rate conversion and display estimated token price for inputAsset.
 - Fixed fiat token prices.
-
 
 ## [1.1.4]